Moves IElevator interface
